+++ packages/components/src/Icon/Icon.stories.tsx
@@ -1,4 +1,4 @@
-import * as React from "react"
+import React from "react"
 import type { ComponentMeta, ComponentStory } from "@storybook/react"
 import { colorMap } from "@greenlabs/formula-design-token"
 import * as IconComponents from "./generated"
@@ -15,14 +15,12 @@ interface IconItemProps {
 // Todo - support code copy(ts, res) handler
 const IconItem: React.FC<IconItemProps> = ({ name, children }) => {
   return (
-    <ThemeScope
-      props={{
-        style: {
-          display: "flex",
-          justifyContent: "center",
-          alignItems: "center",
-          flexDirection: "column",
-        },
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        flexDirection: "column",
       }}
     >
       <div
@@ -40,7 +38,7 @@ const IconItem: React.FC<IconItemProps> = ({ name, children }) => {
         {children}
       </div>
       <div>{name}</div>
-    </ThemeScope>
+    </div>
   )
 }
 
@@ -50,13 +48,13 @@ export const IconGallery: ComponentStory<typeof IconComponents["AppleFill"]> = (
   args
 ) => {
   return (
-    <ThemeScope>
+    <>
       {Object.entries(IconComponents).map(([name, Component]) => (
         <IconItem name={name}>
           <Component {...args} />
         </IconItem>
       ))}
-    </ThemeScope>
+    </>
   )
 }
 
@@ -79,19 +77,21 @@ export default {
   },
   decorators: [
     (Story) => (
-      <div
-        style={{
-          marginTop: "20px",
-          marginBottom: "20px",
-          display: "grid",
-          gridTemplateColumns: "repeat(auto-fill, minmax(200px, 1fr))",
-          overflow: "hidden",
-          flexFlow: "wrap",
-          gap: 30,
+      <ThemeScope
+        props={{
+          style: {
+            marginTop: "20px",
+            marginBottom: "20px",
+            display: "grid",
+            gridTemplateColumns: "repeat(auto-fill, minmax(200px, 1fr))",
+            overflow: "hidden",
+            flexFlow: "wrap",
+            gap: 30,
+          },
         }}
       >
         <Story />
-      </div>
+      </ThemeScope>
     ),
   ],
 } as ComponentMeta<typeof IconComponents["AppleFill"]>